+++ CHANGELOG.md
@@ -21,3 +21,13 @@
 ## 0.0.3+1
 
 * Bug fix from bug fix
+
+## 1.0.0
+
+* Production ready
+* Added `logging` for logging (you should configure logging in your app)
+* Added some hooks for widgets initialization
+* Simplified `ViewWidget`
+* Fix a bug when registering the same dependency twice would trigger the circular dependency error (now it will throw an exception)
+* Improved initialization (parallel initialization, when it is possible)
+* Added some tests